app/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
gunicorn==20.1.0
